add observation and forecast logic to CloneForm view, refactor some common functions into BaseView
