+++ sfa_dash/blueprints/base.py
@@ -6,8 +6,9 @@ from flask import url_for, render_template, request, flash, current_app
 from flask.views import MethodView
 import pandas as pd
 
-from sfa_dash.errors import DataRequestException
+from sfa_dash.api_interface import sites, aggregates
 from sfa_dash.blueprints.util import timeseries_adapter
+from sfa_dash.errors import DataRequestException
 
 
 class BaseView(MethodView):
@@ -248,6 +249,75 @@ class BaseView(MethodView):
                     new_dict[key] = _pop_nonjson(val)
             return new_dict
         return _pop_nonjson(self.metadata)
+
+    def set_site_or_aggregate_metadata(self):
+        """Searches for a site_id or aggregate_id  in self.metadata
+        and loads the expected metadata object from the api in either
+        the 'site' or 'aggregate' key. If the object could not be retrieved,
+        sets a warning, sets the 'plot' template argument to None and
+        reraises the DataRequestError.
+        """
+        if self.metadata.get('site_id') is not None:
+            try:
+                self.metadata['site'] = sites.get_metadata(
+                    self.metadata['site_id'])
+            except DataRequestException:
+                self.temp_args.update({
+                    'warnings': {
+                        'Site Access': [
+                            'Site inaccessible. Plots will not be displayed.']
+                    },
+                })
+                raise
+        elif self.metadata.get('aggregate_id'):
+            try:
+                self.metadata['aggregate'] = aggregates.get_metadata(
+                    self.metadata['aggregate_id'])
+            except DataRequestException:
+                self.temp_args.update({
+                    'warnings': {
+                        'Aggregate Access': [
+                            'Aggregate inaccessible. Plots will not be '
+                            'displayed.']
+                    },
+                })
+                raise
+
+    def set_site_or_aggregate_link(self):
+        """Creates a url to the forecast's site or aggregate. For injection
+        into the metadata template.
+
+        Returns
+        -------
+        link_html: str
+            An anchor tag referencing the site/aggregates page on the
+            dashboard or a string uuid if the metadata wasn't not available,
+            indicating that the current user does not have access to the
+            site or aggregate.
+        """
+        if self.metadata.get('site_id') is not None:
+            if self.metadata.get('site') is not None:
+                site_metadata = self.metadata.get('site')
+                site_name = site_metadata['name']
+                site_id = site_metadata['site_id']
+                site_href = url_for('data_dashboard.site_view',
+                                    uuid=site_id)
+                link_html = f'<a href="{site_href}">{site_name}</a>'
+            else:
+                link_html = self.metadata['site_id']
+        elif self.metadata.get('aggregate_id') is not None:
+            if self.metadata.get('aggregate') is not None:
+                aggregate_metadata = self.metadata.get('aggregate')
+                aggregate_name = aggregate_metadata['name']
+                aggregate_id = aggregate_metadata['aggregate_id']
+                aggregate_href = url_for('data_dashboard.aggregate_view',
+                                         uuid=aggregate_id)
+                link_html = f'<a href="{aggregate_href}">{aggregate_name}</a>'
+            else:
+                link_html = self.metadata['aggregate_id']
+        else:
+            link_html = 'Object Deleted'
+        self.metadata['location_link'] = link_html
 
     def template_args(self):
         return {}
